Change brightness from int to double
I found that brightness is a floating point number for my devices and the app through an exception because it couldn't convert from double to int

+++ lib/src/models/entity.dart
@@ -47,7 +47,7 @@ class EntityAttributes {
 
   // Light
   List<String>? options, supportedColorModes;
-  int? brightness;
+  double? brightness;
   List<int>? rgbColor;
 
   // Climate
